+++ src/main/java/com/example/tripKo/_core/security/filter/RefreshTokenFilter.java
@@ -9,6 +9,7 @@ import com.example.tripKo._core.security.data.JwtValidationType;
 import com.example.tripKo._core.security.data.TokenValidationDTO;
 import com.example.tripKo._core.utils.RedisUtil;
 import java.io.IOException;
+import java.util.Optional;
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.ServletRequest;
@@ -24,39 +25,35 @@ import org.springframework.web.filter.GenericFilterBean;
 @RequiredArgsConstructor
 public class RefreshTokenFilter extends GenericFilterBean {
 
-//    private final JwtProvider jwtProvider;
-//    private final RedisUtil redisUtil;
-//
+    private final JwtProvider jwtProvider;
+    private final RedisUtil redisUtil;
+
     @Override
     public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
             throws IOException, ServletException {
-//        HttpServletRequest httpServletRequest = (HttpServletRequest) request;
-//        final var accessToken = jwtProvider.validateToken(httpServletRequest, ACCESS_TOKEN);
-//        final var refreshToken = jwtProvider.validateToken(httpServletRequest, REFRESH_TOKEN);
-//        boolean isAccessExpiredAndRefreshValid =
-//                isAccessTokenExpired(accessToken.getJwtValidationType()) && isRefreshTokenValid(refreshToken)
-//                        && isTokenInRedis(refreshToken);
-//
-//        if (isAccessExpiredAndRefreshValid) {
-//            Authentication authentication = jwtProvider.getAuthentication(refreshToken.getToken());
-//            SecurityContextHolder.getContext().setAuthentication(authentication);
-//        }
-//
+        HttpServletRequest httpServletRequest = (HttpServletRequest) request;
+        String accessToken = jwtProvider.resolveAccessToken(httpServletRequest);
+
+        if (accessToken != null) {
+        String id = jwtProvider.getAuthId(accessToken);
+
+        final var refreshToken = jwtProvider.resolveRefreshToken(httpServletRequest);
+        boolean isRefreshTokenValid = jwtProvider.validateToken(refreshToken);
+
+        boolean isAccessExpiredAndRefreshValid = isRefreshTokenValid && isTokenInRedis(id, refreshToken);
+
+        if (isAccessExpiredAndRefreshValid) {
+            String newAccessToken = jwtProvider.createAccessToken(id);
+            Authentication authentication = jwtProvider.getAuthentication(newAccessToken);
+            SecurityContextHolder.getContext().setAuthentication(authentication);
+            }
+        }
         chain.doFilter(request, response);
     }
-//
-//    private static boolean isAccessTokenExpired(JwtValidationType jwtValidationType) {
-//        return EXPIRED.equals(jwtValidationType);
-//    }
-//
-//    private static boolean isRefreshTokenValid(TokenValidationDTO tokenValidationDTO) {
-//        return tokenValidationDTO.isValid();
-//    }
-//
-//    private boolean isTokenInRedis(TokenValidationDTO tokenValidationDTO) {
-//        long id = jwtProvider.getAuthId(tokenValidationDTO.getToken());
-//        String tokenInRedis = redisUtil.getData(String.valueOf(id));
-//        return tokenInRedis.equals(tokenValidationDTO.getToken());
-//    }
+
+    private boolean isTokenInRedis(String id, String refreshToken) {
+        Optional<String> tokenInRedis = redisUtil.getData(id);
+        return tokenInRedis.orElseThrow(() -> new RuntimeException("Token not found in Redis")).equals(refreshToken);
+    }
 
 }
